rag_opensearch/requirements.txt
numpy==2.3.5
tqdm==4.67.1
python-dotenv==1.2.1
opensearch-py==3.1.0
langchain-core==1.0.7
langchain-text-splitters==1.0.0
google-genai==1.52.0
pdfplumber==0.11.8
pytesseract==0.3.13
